Rename box to enclosure for consistency
