+++ src/enclosure/enclosure.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 */
 
-#include "box.h"
+#include "enclosure.hpp"
 
 #include "common.h"
 #include "xml_parser.hpp"
