+++ src/enclosure/enclosure_editor.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 */
 
-#include "boxeditor.h"
+#include "enclosure_editor.hpp"
 
 #include "common.h"
 #include "plot.hpp"
