+++ src/enclosure/enclosure_history.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 */
 
-#include "boxhistory.h"
+#include "enclosure_history.hpp"
 
 #include "gspeakersfilechooser.h"
