+++ src/enclosure/enclosure_list.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 */
 
-#include "boxlist.h"
+#include "enclosure_list.hpp"
 #include "common.h"
 
 #include <libxml/parser.h>
